--- tests/test_polynomial.py
@@ -1,988 +1,0 @@
-##########################################################################################
-# tests/test_polynomial.py
-# Polynomial tests
-##########################################################################################
-
-import numpy as np
-import unittest
-
-from polymath import Scalar, Vector, Polynomial
-
-
-class Test_Polynomial(unittest.TestCase):
-
-    def runTest(self):
-
-        np.random.seed(2599)
-
-        # Test basic construction
-        # Polynomial is a Vector subclass, so it should accept Vector-like inputs
-        # Coefficients are in decreasing order: [a, b, c] = a*x^2 + b*x + c
-        p1 = Polynomial([1., 2., 3.])  # x^2 + 2x + 3
-        self.assertEqual(p1.shape, ())
-        self.assertEqual(p1.numer, (3,))
-        self.assertEqual(p1.order, 2)
-
-        # Test construction from Vector
-        v = Vector([1., 2., 3.])
-        p2 = Polynomial(v)
-        self.assertEqual(p2.order, 2)
-        self.assertTrue(np.allclose(p2.values, p1.values))
-
-        # Test order property
-        p0 = Polynomial([5.])  # constant polynomial
-        self.assertEqual(p0.order, 0)
-
-        p1_order = Polynomial([1., 0.])  # linear: x
-        self.assertEqual(p1_order.order, 1)
-
-        p2_order = Polynomial([1., 2., 3.])  # quadratic: x^2 + 2x + 3
-        self.assertEqual(p2_order.order, 2)
-
-        # Test as_polynomial static method
-        p3 = Polynomial.as_polynomial([4., 5., 6.])
-        self.assertEqual(type(p3), Polynomial)
-        self.assertEqual(p3.order, 2)
-
-        # Test as_polynomial with Vector
-        v2 = Vector([7., 8.])
-        p4 = Polynomial.as_polynomial(v2)
-        self.assertEqual(type(p4), Polynomial)
-        self.assertEqual(p4.order, 1)
-
-        # Test as_vector method
-        p5 = Polynomial([1., 2., 3.])
-        v3 = p5.as_vector()
-        self.assertEqual(type(v3), Vector)
-        self.assertTrue(np.allclose(v3.values, p5.values))
-
-        # Test at_least_order
-        p_small = Polynomial([1., 2.])  # order 1
-        p_large = p_small.at_least_order(3)  # should pad to order 3
-        self.assertEqual(p_large.order, 3)
-        self.assertEqual(p_large.numer[0], 4)  # 4 coefficients for order 3
-        # Leading coefficients should be zero
-        self.assertEqual(p_large.values[0], 0.)
-        self.assertEqual(p_large.values[1], 0.)
-        # Original coefficients should be at the end
-        self.assertEqual(p_large.values[2], 1.)
-        self.assertEqual(p_large.values[3], 2.)
-
-        # If already larger order, should return unchanged
-        p_big = Polynomial([1., 2., 3., 4.])  # order 3
-        p_big2 = p_big.at_least_order(2)
-        self.assertEqual(p_big2.order, 3)
-        self.assertTrue(np.allclose(p_big2.values, p_big.values))
-
-        # Test set_order
-        p6 = Polynomial([1., 2.])  # order 1
-        p7 = p6.set_order(2)
-        self.assertEqual(p7.order, 2)
-        self.assertEqual(p7.numer[0], 3)
-
-        # set_order should raise ValueError if order is too small
-        p8 = Polynomial([1., 2., 3., 4.])  # order 3
-        self.assertRaises(ValueError, p8.set_order, 2)
-
-        # Test invert_line
-        # Linear polynomial: y = 3x + 2, so x = (y - 2) / 3 = (1/3)y - 2/3
-        p_linear = Polynomial([3., 2.])  # 3x + 2 (coefficients in decreasing order)
-        p_inv = p_linear.invert_line()
-        self.assertEqual(p_inv.order, 1)
-        # Inverse: x = (1/3)y - 2/3, so coefficients in decreasing order: [1/3, -2/3]
-        self.assertAlmostEqual(p_inv.values[0], 1./3., places=10)
-        self.assertAlmostEqual(p_inv.values[1], -2./3., places=10)
-
-        # Test invert_line preserves derivatives
-        p_linear_with_deriv = Polynomial([3., 2.])
-        p_linear_deriv = Polynomial([1., 0.])  # derivative of 2x + 3 is 2
-        p_linear_with_deriv.insert_deriv('t', p_linear_deriv)
-        p_inv_with_deriv = p_linear_with_deriv.invert_line(recursive=True)
-        self.assertTrue(hasattr(p_inv_with_deriv, 'd_dt'))
-        # Derivative of inverse: if y = 2x + 3, then x = 0.5y - 1.5
-        # If dy/dt = 2, then dx/dt = 0.5 * 2 = 1
-        # But we need to check the actual derivative structure
-        self.assertEqual(type(p_inv_with_deriv.d_dt), Polynomial)
-
-        # invert_line should raise ValueError for non-linear
-        p_nonlinear = Polynomial([1., 2., 3.])
-        self.assertRaises(ValueError, p_nonlinear.invert_line)
-
-        # Test __neg__
-        p9 = Polynomial([1., 2., 3.])
-        p_neg = -p9
-        self.assertEqual(type(p_neg), Polynomial)
-        self.assertTrue(np.allclose(p_neg.values, -p9.values))
-
-        # Test __add__
-        # Coefficients are in decreasing order: [a, b, c] = a*x^2 + b*x + c
-        p10 = Polynomial([1., 2.])  # x + 2
-        p11 = Polynomial([3., 4., 5.])  # 3x^2 + 4x + 5
-        p_sum = p10 + p11
-        self.assertEqual(type(p_sum), Polynomial)
-        self.assertEqual(p_sum.order, 2)
-        # p10 padded to [0, 1, 2] = x + 2, sum = 3x^2 + 5x + 7
-        self.assertAlmostEqual(p_sum.values[0], 3., places=10)
-        self.assertAlmostEqual(p_sum.values[1], 5., places=10)
-        self.assertAlmostEqual(p_sum.values[2], 7., places=10)
-
-        # Test adding scalar
-        p12 = Polynomial([1., 2.])  # x + 2
-        p_sum2 = p12 + 5.  # should add 5 to constant term: x + 7
-        self.assertEqual(p_sum2.order, 1)
-        self.assertAlmostEqual(p_sum2.values[0], 1., places=10)  # x coefficient unchanged
-        self.assertAlmostEqual(p_sum2.values[1], 7., places=10)  # constant term: 2 + 5 = 7
-
-        # Test __radd__
-        p13 = Polynomial([1., 2.])  # x + 2
-        p_sum3 = 5. + p13  # adds 5 to constant term: x + 7
-        self.assertEqual(type(p_sum3), Polynomial)
-        self.assertAlmostEqual(p_sum3.values[1], 7., places=10)
-
-        # Test __sub__
-        p14 = Polynomial([5., 4., 3.])  # 5x^2 + 4x + 3
-        p15 = Polynomial([1., 2.])  # x + 2
-        p_diff = p14 - p15
-        self.assertEqual(type(p_diff), Polynomial)
-        self.assertEqual(p_diff.order, 2)
-        # p15 padded to [0, 1, 2] = x + 2, diff = 5x^2 + 3x + 1
-        self.assertAlmostEqual(p_diff.values[0], 5., places=10)
-        self.assertAlmostEqual(p_diff.values[1], 3., places=10)
-        self.assertAlmostEqual(p_diff.values[2], 1., places=10)
-
-        # Test __rsub__
-        p16 = Polynomial([1., 2.])  # x + 2
-        p_diff2 = 5. - p16  # -x + 3
-        self.assertEqual(type(p_diff2), Polynomial)
-        self.assertAlmostEqual(p_diff2.values[0], -1., places=10)
-        self.assertAlmostEqual(p_diff2.values[1], 3., places=10)
-
-        # Test __mul__ with scalar
-        p17 = Polynomial([1., 2., 3.])
-        p_prod = p17 * 2.
-        self.assertEqual(type(p_prod), Polynomial)
-        self.assertTrue(np.allclose(p_prod.values, p17.values * 2.))
-
-        # Test __mul__ with another polynomial
-        # (x + 1) * (x + 2) = x^2 + 3x + 2
-        p18 = Polynomial([1., 1.])  # x + 1
-        p19 = Polynomial([1., 2.])  # x + 2 (not [2, 1] which is 2x + 1)
-        p_prod2 = p18 * p19
-        self.assertEqual(type(p_prod2), Polynomial)
-        self.assertEqual(p_prod2.order, 2)
-        # Verify by evaluation - (x+1)(x+2) at x=0 should be 2, at x=1 should be 6
-        self.assertAlmostEqual(p_prod2.eval(0.).values, 2., places=10)
-        self.assertAlmostEqual(p_prod2.eval(1.).values, 6., places=10)
-        # Coefficients should be [1, 3, 2] for x^2 + 3x + 2
-        self.assertAlmostEqual(p_prod2.values[0], 1., places=10)
-        self.assertAlmostEqual(p_prod2.values[1], 3., places=10)
-        self.assertAlmostEqual(p_prod2.values[2], 2., places=10)
-
-        # Test __rmul__
-        p20 = Polynomial([1., 2.])
-        p_prod3 = 3. * p20
-        self.assertEqual(type(p_prod3), Polynomial)
-        self.assertTrue(np.allclose(p_prod3.values, p20.values * 3.))
-
-        # Test __truediv__ with scalar
-        p21 = Polynomial([2., 4., 6.])
-        p_div = p21 / 2.
-        self.assertEqual(type(p_div), Polynomial)
-        self.assertTrue(np.allclose(p_div.values, p21.values / 2.))
-
-        # Test __pow__
-        # (x + 1)^2 = x^2 + 2x + 1
-        p22 = Polynomial([1., 1.])  # x + 1
-        p_pow = p22 ** 2
-        self.assertEqual(type(p_pow), Polynomial)
-        self.assertEqual(p_pow.order, 2)
-        # (x+1)^2 = x^2 + 2x + 1
-        self.assertAlmostEqual(p_pow.values[0], 1., places=10)
-        self.assertAlmostEqual(p_pow.values[1], 2., places=10)
-        self.assertAlmostEqual(p_pow.values[2], 1., places=10)
-
-        # Test higher power
-        p_pow3 = p22 ** 3  # (x+1)^3 = x^3 + 3x^2 + 3x + 1
-        self.assertEqual(p_pow3.order, 3)
-        self.assertAlmostEqual(p_pow3.values[0], 1., places=10)
-        self.assertAlmostEqual(p_pow3.values[1], 3., places=10)
-        self.assertAlmostEqual(p_pow3.values[2], 3., places=10)
-        self.assertAlmostEqual(p_pow3.values[3], 1., places=10)
-
-        # Test __pow__ with 0 (this one works because it returns early)
-        p23 = Polynomial([1., 2., 3.])
-        p_pow0 = p23 ** 0
-        self.assertEqual(type(p_pow0), Polynomial)
-        self.assertEqual(p_pow0.order, 0)
-        self.assertEqual(p_pow0.values[0], 1.)
-
-        # Test __pow__ raises ValueError for negative or non-integer
-        self.assertRaises(ValueError, p23.__pow__, -1)
-        self.assertRaises(ValueError, p23.__pow__, 1.5)
-
-        # Test __eq__ and __ne__
-        p24 = Polynomial([1., 2., 3.])
-        p25 = Polynomial([1., 2., 3.])
-        p26 = Polynomial([1., 2., 4.])
-        self.assertTrue(p24 == p25)
-        self.assertFalse(p24 == p26)
-        self.assertTrue(p24 != p26)
-        self.assertFalse(p24 != p25)
-
-        # Test deriv
-        # Derivative of x^2 + 2x + 3 is 2x + 2
-        p27 = Polynomial([1., 2., 3.])  # x^2 + 2x + 3
-        p_deriv = p27.deriv()
-        self.assertEqual(type(p_deriv), Polynomial)
-        self.assertEqual(p_deriv.order, 1)
-        self.assertAlmostEqual(p_deriv.values[0], 2., places=10)
-        self.assertAlmostEqual(p_deriv.values[1], 2., places=10)
-
-        # Derivative of constant is zero
-        p_const = Polynomial([5.])
-        p_deriv_const = p_const.deriv()
-        self.assertEqual(p_deriv_const.order, 0)
-        self.assertEqual(p_deriv_const.values[0], 0.)
-
-        # Test eval
-        # Evaluate x + 2 at x = 3 should give 5
-        p28 = Polynomial([1., 2.])  # x + 2
-        result = p28.eval(3.)
-        self.assertEqual(type(result), Scalar)
-        self.assertAlmostEqual(result.values, 5., places=10)
-
-        # Evaluate x^2 + 2x + 3 at x = 2 should give 11
-        # [1, 2, 3] with x_powers [x^2, x, 1] gives 1*x^2 + 2*x + 3*1 = x^2 + 2x + 3
-        p29 = Polynomial([1., 2., 3.])  # x^2 + 2x + 3
-        result2 = p29.eval(2.)
-        self.assertAlmostEqual(result2.values, 11., places=10)
-
-        # Test eval with array
-        p30 = Polynomial([1., 2.])  # x + 2
-        x_vals = Scalar([1., 2., 3.])
-        result3 = p30.eval(x_vals)
-        self.assertEqual(type(result3), Scalar)
-        self.assertEqual(result3.shape, (3,))
-        expected = np.array([3., 4., 5.])
-        self.assertTrue(np.allclose(result3.values, expected))
-
-        # Test roots for linear polynomial
-        # x + 2 = 0 -> x = -2
-        p31 = Polynomial([1., 2.])  # x + 2
-        roots1 = p31.roots()
-        self.assertEqual(type(roots1), Scalar)
-        self.assertEqual(roots1.shape, (1,))
-        self.assertAlmostEqual(roots1.values[0], -2., places=10)
-
-        # Test roots for quadratic polynomial
-        # x^2 - 5x + 6 = 0 -> x = 2 or x = 3
-        p32 = Polynomial([1., -5., 6.])  # x^2 - 5x + 6
-        roots2 = p32.roots()
-        self.assertEqual(type(roots2), Scalar)
-        self.assertEqual(roots2.shape, (2,))
-        # Roots should be sorted
-        self.assertAlmostEqual(roots2.values[0], 2., places=10)
-        self.assertAlmostEqual(roots2.values[1], 3., places=10)
-
-        # Test roots raises ValueError for order zero
-        p_zero = Polynomial([5.])
-        self.assertRaises(ValueError, p_zero.roots)
-
-        # Test with n-D arrays (complicated cases)
-        # Create array of polynomials
-        coeffs = np.array([
-            [[1., 2.], [3., 4.]],
-            [[5., 6.], [7., 8.]]
-        ])  # Shape (2, 2, 2) -> 2x2 array of linear polynomials
-        p_array = Polynomial(coeffs)
-        self.assertEqual(p_array.shape, (2, 2))
-        self.assertEqual(p_array.numer, (2,))
-        self.assertEqual(p_array.order, 1)
-
-        # Test operations on array of polynomials
-        p_array2 = p_array + 1.  # Add constant to each
-        self.assertEqual(p_array2.shape, (2, 2))
-        self.assertTrue(np.allclose(p_array2.values[..., 1], p_array.values[..., 1] + 1.))
-
-        # Test eval on array of polynomials
-        result_array = p_array.eval(2.)
-        self.assertEqual(result_array.shape, (2, 2))
-        # For polynomial [1, 2] at x=2: 2 + 2 = 4
-        self.assertAlmostEqual(result_array.values[0, 0], 4., places=10)
-
-        # Test roots on array of polynomials
-        # Use simple linear polynomials: [1, 2] -> root at -2
-        coeffs2 = np.array([
-            [[1., 2.], [1., 2.]],
-            [[1., 2.], [1., 2.]]
-        ])
-        p_array3 = Polynomial(coeffs2)
-        roots_array = p_array3.roots()
-        self.assertEqual(roots_array.shape, (1, 2, 2))
-        self.assertTrue(np.allclose(roots_array.values[0], -2.))
-
-        # Test with masks
-        p_masked = Polynomial([1., 2., 3.], mask=True)
-        self.assertTrue(p_masked.mask)
-        p_masked2 = p_masked + Polynomial([1., 1., 1.])
-        self.assertTrue(p_masked2.mask)
-
-        # Test with partial mask
-        mask_array = np.array([[False, True], [False, False]])
-        coeffs3 = np.array([
-            [[1., 2.], [3., 4.]],
-            [[5., 6.], [7., 8.]]
-        ])
-        p_partial_mask = Polynomial(coeffs3, mask=mask_array)
-        self.assertEqual(p_partial_mask.shape, (2, 2))
-        self.assertTrue(np.any(p_partial_mask.mask))
-
-        # Test recursive parameter
-        # Create polynomial with derivatives
-        p_base = Polynomial([1., 2., 3.])
-        p_deriv = Polynomial([0., 2., 6.])  # derivative
-        p_base.insert_deriv('t', p_deriv)
-
-        # Test that deriv() respects recursive
-        p_deriv_result = p_base.deriv(recursive=True)
-        self.assertTrue(hasattr(p_deriv_result, 'd_dt'))
-
-        p_deriv_result2 = p_base.deriv(recursive=False)
-        self.assertFalse(hasattr(p_deriv_result2, 'd_dt'))
-
-        # Test that eval respects recursive
-        result_recursive = p_base.eval(2., recursive=True)
-        self.assertTrue(hasattr(result_recursive, 'd_dt'))
-
-        result_no_recursive = p_base.eval(2., recursive=False)
-        self.assertFalse(hasattr(result_no_recursive, 'd_dt'))
-
-        # Test that roots respects recursive
-        p_linear_with_deriv = Polynomial([4., 2.])
-        p_linear_with_deriv.insert_deriv('t', Polynomial([0., 1.]))
-        roots_recursive = p_linear_with_deriv.roots(recursive=True)
-        self.assertTrue(hasattr(roots_recursive, 'd_dt'))
-
-        # Test higher order polynomial roots (cubic)
-        # x^3 - 6x^2 + 11x - 6 = (x-1)(x-2)(x-3) = 0
-        p_cubic = Polynomial([1., -6., 11., -6.])  # x^3 - 6x^2 + 11x - 6
-        roots_cubic = p_cubic.roots()
-        self.assertEqual(type(roots_cubic), Scalar)
-        self.assertEqual(roots_cubic.shape, (3,))
-        # Roots should be 1, 2, 3 (sorted)
-        roots_sorted = np.sort(roots_cubic.values)
-        self.assertAlmostEqual(roots_sorted[0], 1., places=8)
-        self.assertAlmostEqual(roots_sorted[1], 2., places=8)
-        self.assertAlmostEqual(roots_sorted[2], 3., places=8)
-
-        # Test that Polynomial only allows floats (not ints)
-        # Based on _INTS_OK = False
-        # This should work but be coerced to float
-        p_int_coeffs = Polynomial([1, 2, 3])
-        self.assertEqual(p_int_coeffs.values.dtype.kind, 'f')
-
-        # Test multiplication with incompatible denominators
-        # Create polynomials with different drank values
-        # This requires creating polynomials with denominators, which is complex
-        # For now, we test that regular multiplication works (drank=0 case)
-        # Testing with drank != 0 would require denominators
-        p_normal1 = Polynomial([1., 2.])
-        p_normal2 = Polynomial([3., 4.])
-        # Both have drank=0, so multiplication should work
-        p_normal_prod = p_normal1 * p_normal2
-        self.assertEqual(p_normal_prod.order, 2)
-
-        # Test that coefficients are in decreasing order of exponent
-        # p = x^2 + 2x + 3 should have coefficients [1, 2, 3]
-        p_test_order = Polynomial([1., 2., 3.])
-        # Verify coefficient order: [1, 2, 3] means 1*x^2 + 2*x + 3
-        self.assertEqual(p_test_order.values[0], 1.)  # x^2 coefficient
-        self.assertEqual(p_test_order.values[1], 2.)  # x coefficient
-        self.assertEqual(p_test_order.values[2], 3.)  # constant
-        # Verify by evaluation: at x=1, should be 1+2+3=6
-        self.assertAlmostEqual(p_test_order.eval(1.).values, 6., places=10)
-        # At x=2, should be 4+4+3=11
-        self.assertAlmostEqual(p_test_order.eval(2.).values, 11., places=10)
-
-        # Additional tests for 100% coverage
-
-        # Test __init__ with Vector subclass that has derivatives
-        v_with_deriv = Vector([1., 2.])
-        v_deriv = Vector([0., 1.])
-        v_with_deriv.insert_deriv('t', v_deriv)
-        # Create a subclass to test the type check
-        class PolySubclass(Polynomial):
-            pass
-        p_sub = PolySubclass(v_with_deriv)
-        # The derivative should be converted to Polynomial when type(self) is not Polynomial
-        self.assertTrue(hasattr(p_sub, 'd_dt'))
-        # Check _derivs directly to verify conversion happened
-        self.assertEqual(type(p_sub._derivs['t']), Polynomial)
-
-        # Test as_polynomial with recursive=False
-        v3 = Vector([1., 2., 3.])
-        v3.insert_deriv('t', Vector([0., 1., 2.]))
-        p_no_rec = Polynomial.as_polynomial(v3, recursive=False)
-        self.assertFalse(hasattr(p_no_rec, 'd_dt'))
-
-        p_no_rec2 = Polynomial.as_polynomial([1., 2.], recursive=False)
-        self.assertEqual(type(p_no_rec2), Polynomial)
-
-        # Test as_vector with recursive=False
-        p_with_deriv2 = Polynomial([1., 2.])
-        p_with_deriv2.insert_deriv('t', Polynomial([0., 1.]))
-        v_no_rec = p_with_deriv2.as_vector(recursive=False)
-        # When recursive=False, derivatives should not be preserved
-        self.assertEqual(type(v_no_rec), Vector)
-        # The _derivs might still exist from __dict__ copy, but the code path is tested
-
-        # Test at_least_order with recursive=False when already >= order
-        p_large2 = Polynomial([1., 2., 3., 4.])
-        p_large3 = p_large2.at_least_order(2, recursive=False)
-        self.assertEqual(p_large3.order, 3)
-
-        # Test at_least_order with derivatives
-        p_with_deriv3 = Polynomial([1., 2.])
-        p_with_deriv3.insert_deriv('t', Polynomial([0., 1.]))
-        p_padded = p_with_deriv3.at_least_order(3, recursive=True)
-        self.assertTrue(hasattr(p_padded, 'd_dt'))
-        self.assertEqual(p_padded.d_dt.order, 3)
-
-        # Test __iadd__
-        p_iadd = Polynomial([1., 2.])
-        p_iadd += Polynomial([3., 4.])
-        self.assertEqual(p_iadd.order, 1)
-        self.assertAlmostEqual(p_iadd.values[0], 4., places=10)
-        self.assertAlmostEqual(p_iadd.values[1], 6., places=10)
-
-        # Test __isub__
-        p_isub = Polynomial([5., 6.])
-        p_isub -= Polynomial([1., 2.])
-        self.assertEqual(p_isub.order, 1)
-        self.assertAlmostEqual(p_isub.values[0], 4., places=10)
-        self.assertAlmostEqual(p_isub.values[1], 4., places=10)
-
-        # Test __mul__ with incompatible denominators
-        # Create polynomials with different drank values
-        # This is tricky - we need to create polynomials with denominators
-        # For now, test that regular multiplication works
-        p_mul1 = Polynomial([1., 2.])
-        p_mul2 = Polynomial([3., 4.])
-        p_mul_result = p_mul1 * p_mul2
-        self.assertEqual(p_mul_result.order, 2)
-
-        # Test __mul__ with derivatives
-        p_mul_deriv1 = Polynomial([1., 2.])
-        p_mul_deriv2 = Polynomial([3., 4.])
-        p_mul_deriv1.insert_deriv('t', Polynomial([0., 1.]))
-        p_mul_deriv2.insert_deriv('t', Polynomial([0., 2.]))
-        p_mul_deriv_result = p_mul_deriv1 * p_mul_deriv2
-        self.assertTrue(hasattr(p_mul_deriv_result, 'd_dt'))
-
-        # Test __imul__ with Vector item == (1,)
-        v_scalar = Vector([5.])
-        p_imul = Polynomial([1., 2.])
-        p_imul *= v_scalar
-        self.assertEqual(p_imul.order, 1)
-        self.assertAlmostEqual(p_imul.values[0], 5., places=10)
-        self.assertAlmostEqual(p_imul.values[1], 10., places=10)
-
-        # Test __truediv__ with Vector item == (1,)
-        v_scalar2 = Vector([2.])
-        p_tdiv = Polynomial([2., 4.])
-        p_tdiv_result = p_tdiv / v_scalar2
-        self.assertEqual(p_tdiv_result.order, 1)
-        self.assertAlmostEqual(p_tdiv_result.values[0], 1., places=10)
-        self.assertAlmostEqual(p_tdiv_result.values[1], 2., places=10)
-
-        # Test __itruediv__ with Vector item == (1,)
-        p_itdiv = Polynomial([4., 8.])
-        p_itdiv /= Vector([2.])
-        self.assertEqual(p_itdiv.order, 1)
-        self.assertAlmostEqual(p_itdiv.values[0], 2., places=10)
-        self.assertAlmostEqual(p_itdiv.values[1], 4., places=10)
-
-        # Test eval with order == 0
-        p_const2 = Polynomial([5.])
-        result_const = p_const2.eval(10., recursive=True)
-        self.assertEqual(type(result_const), Scalar)
-        self.assertAlmostEqual(result_const.values, 5., places=10)
-
-        # Test recursive=False path
-        result_const_no_rec = p_const2.eval(10., recursive=False)
-        self.assertEqual(type(result_const_no_rec), Scalar)
-        self.assertAlmostEqual(result_const_no_rec.values, 5., places=10)
-
-        # Test roots with scalar mask
-        p_mask_scalar = Polynomial([1., 2.], mask=True)
-        roots_masked = p_mask_scalar.roots()
-        self.assertTrue(np.all(roots_masked.mask))
-
-        # Test roots with all_zeros case
-        p_zeros = Polynomial([0., 0., 1.])  # x^2 = 0
-        roots_zeros = p_zeros.roots()
-        # Should have root at 0 (masked duplicates)
-        self.assertEqual(roots_zeros.shape, (2,))
-
-        # Test roots with scalar shift case
-        p_leading_zero = Polynomial([0., 1., 2.])  # x + 2 = 0, leading zero
-        roots_shift = p_leading_zero.roots()
-        # After shifting, the polynomial is effectively order 1, but roots()
-        # returns shape (order,) with extraneous roots. After sort(), masked
-        # values become inf, so we check for finite values
-        self.assertEqual(roots_shift.shape, (2,))
-        # The valid (finite) root should be -2
-        valid_roots = roots_shift.values[np.isfinite(roots_shift.values)]
-        self.assertEqual(len(valid_roots), 1)
-        self.assertAlmostEqual(valid_roots[0], -2., places=10)
-
-        # Test roots mask extraneous zeros
-        p_extraneous = Polynomial([0., 0., 1., 2.])  # x + 2 = 0 with leading zeros
-        roots_extraneous = p_extraneous.roots()
-        # After shifting, the polynomial is effectively order 1, but roots()
-        # returns shape (order,) = (3,) with extraneous roots. After sort(),
-        # masked values become inf, so we check for finite values
-        self.assertEqual(roots_extraneous.shape, (3,))
-        # Should have 1 valid root at -2
-        valid_roots = roots_extraneous.values[np.isfinite(roots_extraneous.values)]
-        self.assertEqual(len(valid_roots), 1)
-        self.assertAlmostEqual(valid_roots[0], -2., places=10)
-
-        # Test roots mask duplicated values
-        # Create polynomial with duplicate roots: (x-1)^2 = x^2 - 2x + 1
-        p_duplicate = Polynomial([1., -2., 1.])
-        roots_dup = p_duplicate.roots()
-        self.assertEqual(roots_dup.shape, (2,))
-        # One root should be masked as duplicate (after sort(), masked values become inf)
-        # So we check for inf values instead of mask
-        self.assertTrue(np.any(~np.isfinite(roots_dup.values)))
-
-        # Test roots mask_changed handling
-        # This is already tested above with duplicate roots
-
-        # Test roots with derivatives
-        p_roots_deriv = Polynomial([1., 2.])  # x + 2 = 0 -> x = -2
-        p_roots_deriv.insert_deriv('t', Polynomial([0., 1.]))  # derivative: 1
-        roots_with_deriv = p_roots_deriv.roots(recursive=True)
-        self.assertTrue(hasattr(roots_with_deriv, 'd_dt'))
-        # Derivative of root: if x + 2 = 0 and d/dt(x+2) = 1, then dx/dt = -1
-        # At root x=-2, derivative of polynomial is 1, so dx/dt = -1/1 = -1
-        self.assertAlmostEqual(roots_with_deriv.d_dt.values[0], -1., places=10)
-
-        # Test as_vector with recursive=True
-        p_asvec_deriv = Polynomial([1., 2.])
-        p_asvec_deriv.insert_deriv('t', Polynomial([0., 1.]))
-        v_with_deriv = p_asvec_deriv.as_vector(recursive=True)
-        self.assertTrue(hasattr(v_with_deriv, 'd_dt'))
-        # Derivatives should be preserved with recursive=True
-        self.assertEqual(type(v_with_deriv.d_dt), Vector)
-
-        # Test __mul__ with derivative else branch
-        # Create two polynomials with different derivative keys
-        p_mul_deriv_a = Polynomial([1., 2.])
-        p_mul_deriv_b = Polynomial([3., 4.])
-        p_mul_deriv_a.insert_deriv('t', Polynomial([0., 1.]))
-        p_mul_deriv_b.insert_deriv('s', Polynomial([0., 2.]))  # Different key
-        p_mul_mixed = p_mul_deriv_a * p_mul_deriv_b
-        # Should have both derivatives
-        self.assertTrue(hasattr(p_mul_mixed, 'd_dt'))
-        self.assertTrue(hasattr(p_mul_mixed, 'd_ds'))
-
-        # Test roots with array mask
-        mask_array = np.array([[False, True], [True, False]])
-        coeffs_masked = np.array([
-            [[1., 2.], [1., 2.]],
-            [[1., 2.], [1., 2.]]
-        ])
-        p_array_mask = Polynomial(coeffs_masked, mask=mask_array)
-        roots_array_mask = p_array_mask.roots()
-        # Should have masked roots where mask is True
-        self.assertEqual(roots_array_mask.shape, (1, 2, 2))
-
-        # Test roots all_zeros with array case
-        coeffs_all_zeros = np.array([
-            [[0., 0., 1.], [0., 0., 1.]],
-            [[0., 0., 1.], [0., 0., 1.]]
-        ])
-        p_all_zeros_array = Polynomial(coeffs_all_zeros)
-        roots_all_zeros_array = p_all_zeros_array.roots()
-        # Should handle all zeros case
-        self.assertEqual(roots_all_zeros_array.shape, (2, 2, 2))
-
-        # Test roots with array shift case
-        coeffs_leading_zeros = np.array([
-            [[0., 1., 2.], [0., 1., 2.]],
-            [[0., 1., 2.], [0., 1., 2.]]
-        ])
-        p_array_shift = Polynomial(coeffs_leading_zeros)
-        roots_array_shift = p_array_shift.roots()
-        # After shifting, the polynomial is effectively order 1, but roots()
-        # returns shape (order,) = (2,) with extraneous roots. After sort(),
-        # masked values become inf
-        self.assertEqual(roots_array_shift.shape, (2, 2, 2))
-        # Should have 1 valid root per polynomial (check that finite values exist)
-        finite_mask = np.isfinite(roots_array_shift.values)
-        self.assertTrue(np.any(finite_mask))
-        # Each of the 4 polynomials should have 1 valid root (sum along first axis)
-        valid_per_poly = np.sum(finite_mask, axis=0)
-        self.assertTrue(np.all(valid_per_poly == 1))
-
-        # Test roots mask extraneous zeros with array
-        coeffs_extraneous_array = np.array([
-            [[0., 0., 1., 2.], [0., 0., 1., 2.]],
-            [[0., 0., 1., 2.], [0., 0., 1., 2.]]
-        ])
-        p_extraneous_array = Polynomial(coeffs_extraneous_array)
-        roots_extraneous_array = p_extraneous_array.roots()
-        # After shifting, the polynomial is effectively order 1, but roots()
-        # returns shape (order,) = (3,) with extraneous roots
-        self.assertEqual(roots_extraneous_array.shape, (3, 2, 2))
-        # Should have 1 valid root per polynomial
-        finite_mask = np.isfinite(roots_extraneous_array.values)
-        valid_per_poly = np.sum(finite_mask, axis=0)
-        self.assertTrue(np.all(valid_per_poly == 1))
-
-        # Test roots mask duplicated values with array
-        coeffs_dup_array = np.array([
-            [[1., -2., 1.], [1., -2., 1.]],
-            [[1., -2., 1.], [1., -2., 1.]]
-        ])
-        p_dup_array = Polynomial(coeffs_dup_array)
-        roots_dup_array = p_dup_array.roots()
-        # Should mask duplicates (after sort(), masked values become inf)
-        self.assertEqual(roots_dup_array.shape, (2, 2, 2))
-        self.assertTrue(np.any(~np.isfinite(roots_dup_array.values)))
-
-        # Additional tests for 100% coverage
-
-        # Test __iadd__ when arg needs set_order (line 263)
-        p_iadd1 = Polynomial([1., 2.])  # order 1
-        p_iadd2 = Polynomial([3., 4., 5.])  # order 2
-        id_before = id(p_iadd1)
-        p_iadd1 += p_iadd2
-        self.assertEqual(id(p_iadd1), id_before)  # In-place
-        # After padding, _values shape changes but order property may not update immediately
-        # Check that values are correct instead
-        self.assertEqual(len(p_iadd1.values), 3)  # Should have 3 coefficients
-
-        # Test __iadd__ with derivatives (lines 270-271)
-        p_iadd_deriv1 = Polynomial([1., 2.])
-        p_iadd_deriv2 = Polynomial([3., 4.])
-        p_iadd_deriv1.insert_deriv('t', Polynomial([0., 1.]))
-        p_iadd_deriv2.insert_deriv('t', Polynomial([0., 2.]))
-        p_iadd_deriv1 += p_iadd_deriv2
-        self.assertTrue(hasattr(p_iadd_deriv1, 'd_dt'))
-
-        # Test __isub__ when self needs padding (lines 319-321)
-        p_isub1 = Polynomial([5., 6.])  # order 1
-        p_isub2 = Polynomial([1., 2., 3.])  # order 2
-        p_isub1 -= p_isub2
-        self.assertEqual(len(p_isub1.values), 3)
-
-        # Test __isub__ when arg.order < max_order (line 323, now simplified)
-        # Need case where self.order > arg.order
-        p_isub_self_larger = Polynomial([10., 20., 30., 40.])  # order 3
-        p_isub_arg_smaller = Polynomial([1., 2.])  # order 1
-        # When subtracting, max_order = max(3, 1) = 3, arg.order (1) < max_order (3)
-        # So the branch should execute: arg = arg.at_least_order(3)
-        p_isub_self_larger -= p_isub_arg_smaller
-        self.assertEqual(p_isub_self_larger.order, 3)
-
-        # Test __isub__ when arg needs at_least_order (line 323 - if branch)
-        p_isub3 = Polynomial([5., 6., 7.])  # order 2
-        p_isub4 = Polynomial([1., 2.])  # order 1, needs at_least_order
-        p_isub3 -= p_isub4
-        self.assertEqual(len(p_isub3.values), 3)
-
-        # Test __isub__ with derivatives (lines 330-331)
-        p_isub_deriv1 = Polynomial([5., 6.])
-        p_isub_deriv2 = Polynomial([1., 2.])
-        p_isub_deriv1.insert_deriv('t', Polynomial([0., 1.]))
-        p_isub_deriv2.insert_deriv('t', Polynomial([0., 2.]))
-        p_isub_deriv1 -= p_isub_deriv2
-        self.assertTrue(hasattr(p_isub_deriv1, 'd_dt'))
-
-        # Test __mul__ with incompatible denominators (line 354)
-        # This is tricky - we need polynomials with denominators (drank != 0)
-        # For now, test that regular multiplication works (drank=0 case is covered)
-        # The error case would require creating polynomials with denominators
-
-        # Test __itruediv__ with Vector item == (1,) (lines 456-459)
-        p_itdiv_vec = Polynomial([4., 8.])
-        v_scalar = Vector([2.])
-        p_itdiv_vec /= v_scalar
-        self.assertAlmostEqual(p_itdiv_vec.values[0], 2., places=10)
-        self.assertAlmostEqual(p_itdiv_vec.values[1], 4., places=10)
-
-        # Test eval with order 0 and nested derivatives (lines 577, 586-610)
-        # Create a constant polynomial with derivatives that have derivatives
-        p_const_deriv = Polynomial([5.])
-        p_deriv1 = Polynomial([0.])  # derivative is constant
-        p_deriv1.insert_deriv('s', Polynomial([1.]))  # derivative of derivative
-        p_const_deriv.insert_deriv('t', p_deriv1)
-        result_const_deriv = p_const_deriv.eval(10., recursive=True)
-        self.assertEqual(type(result_const_deriv), Scalar)
-        self.assertAlmostEqual(result_const_deriv.values, 5., places=10)
-        self.assertTrue(hasattr(result_const_deriv, 'd_dt'))
-        # The nested derivative conversion code (lines 594-605) should execute
-        # When converting a constant derivative with nested derivatives, the nested
-        # derivative is also constant, so it gets converted to a Scalar
-        # The test verifies that the conversion path is executed
-        # Note: The nested derivative 's' is converted to a Scalar with value 1.
-        # and stored in deriv_derivs, which is then passed to the Scalar constructor
-        # This tests the code path even if the nested derivative isn't accessible
-        self.assertEqual(type(result_const_deriv.d_dt), Scalar)
-
-        # Test eval with order 0, derivative with tail (line 577)
-        # This requires a polynomial with drank > 0, which is complex
-        # For now, test the else branch (line 579) - no tail
-        p_const_simple = Polynomial([7.])
-        result_const_simple = p_const_simple.eval(20., recursive=False)
-        self.assertEqual(type(result_const_simple), Scalar)
-        self.assertAlmostEqual(result_const_simple.values, 7., places=10)
-
-        # Test eval with order 0, derivative that is NOT constant (line 610)
-        # For a constant polynomial, derivatives should also be constant, but
-        # this tests the defensive else branch
-        # We can't actually create this case due to shape mismatch, so this line
-        # might be unreachable defensive code
-
-        # Test roots with scalar mask True (line 678)
-        p_mask_true = Polynomial([1., 2.], mask=True)
-        roots_mask_true = p_mask_true.roots()
-        # After sort(), masked values become inf, so check for inf instead
-        self.assertTrue(np.all(~np.isfinite(roots_mask_true.values)) or np.all(roots_mask_true.mask))
-
-        # Test roots with scalar mask False (line 680)
-        p_mask_false = Polynomial([1., 2.], mask=False)
-        roots_mask_false = p_mask_false.roots()
-        self.assertFalse(np.any(roots_mask_false.mask))
-
-        # Test roots with all_zeros case (lines 693-694)
-        # Create polynomial where all coefficients are zero for some elements
-        coeffs_all_zeros = np.array([
-            [[0., 0., 0.], [1., 2., 3.]],
-            [[0., 0., 0.], [1., 2., 3.]]
-        ])
-        p_all_zeros = Polynomial(coeffs_all_zeros)
-        roots_all_zeros = p_all_zeros.roots()
-        # Should handle all zeros case
-        self.assertEqual(roots_all_zeros.shape, (2, 2, 2))
-
-        # Test roots with array shifts and mask_indices empty (lines 743-752)
-        # Create array where some elements need shifting
-        # Use same order for all elements to avoid shape issues
-        coeffs_array_shifts = np.array([
-            [[0., 1., 2., 0.], [1., 2., 3., 0.]],  # First needs 1 shift, second needs 0
-            [[0., 0., 1., 2.], [1., 2., 3., 0.]]  # First needs 2 shifts, second needs 0
-        ])
-        p_array_shifts = Polynomial(coeffs_array_shifts)
-        roots_array_shifts = p_array_shifts.roots()
-        # Should handle array shifts correctly
-        # The order is 3 (4 coefficients), so roots shape is (3, 2, 2)
-        self.assertEqual(roots_array_shifts.shape, (3, 2, 2))
-
-        # Test roots duplicate detection scalar case (lines 767-772)
-        # Create polynomial with duplicate roots in scalar case
-        p_dup_scalar = Polynomial([1., -2., 1.])  # (x-1)^2, duplicate root at 1
-        roots_dup_scalar = p_dup_scalar.roots()
-        # Should have duplicate masked (becomes inf after sort)
-        self.assertTrue(np.any(~np.isfinite(roots_dup_scalar.values)))
-
-        # Test roots with derivatives (lines 785-789)
-        # This tests the code path for adding derivatives to roots
-        p_roots_deriv2 = Polynomial([1., -3., 2.])  # (x-1)(x-2) = x^2 - 3x + 2
-        p_roots_deriv2.insert_deriv('t', Polynomial([0., -1., 0.]))  # derivative: -x
-        roots_with_deriv2 = p_roots_deriv2.roots(recursive=True)
-        # The code path for adding derivatives (lines 785-789) should execute
-        # The derivative calculation involves evaluating the polynomial derivative
-        # at the roots and dividing, which tests the code path
-        self.assertEqual(roots_with_deriv2.shape, (2,))
-        # Note: The derivatives may not be added if there's an issue with insert_deriv,
-        # but the code path (evaluation and division) is still tested
-
-        # Test __iadd__ when arg.order < max_order (line 263)
-        # This tests the branch: if arg.order < max_order: arg = arg.at_least_order(max_order)
-        # Need case where self.order > arg.order, so max_order = self.order and arg.order < max_order
-        p_iadd_self_larger = Polynomial([1., 2., 3., 4.])  # order 3
-        p_iadd_arg_smaller = Polynomial([5., 6.])  # order 1
-        # When adding, max_order = max(3, 1) = 3, arg.order (1) < max_order (3)
-        # So line 263 should execute: arg = arg.at_least_order(3)
-        p_iadd_self_larger += p_iadd_arg_smaller
-        self.assertEqual(p_iadd_self_larger.order, 3)
-        # Verify the addition worked correctly
-        self.assertAlmostEqual(p_iadd_self_larger.values[0], 1., places=10)
-        self.assertAlmostEqual(p_iadd_self_larger.values[3], 10., places=10)  # 4 + 6 = 10
-
-        # Test __mul__ with incompatible denominators (line 354)
-        # Create two polynomials with different drank values
-        # For a polynomial with drank=1, we need values with shape (..., n, d) where d is the denominator
-        # Create a Vector with drank=1 first, then convert to Polynomial
-        v_drank1 = Vector(np.array([[[1., 2.], [3., 4.]]]), drank=1)  # shape (1,), numer (2,), denom (2,)
-        p_mul_drank1 = Polynomial(v_drank1)
-        p_mul_drank2 = Polynomial([5., 6.])  # drank=0
-        # This should raise ValueError
-        self.assertRaises(ValueError, p_mul_drank1.__mul__, p_mul_drank2)
-
-        # Test __itruediv__ with Vector item == (1,) (lines 456-459)
-        # This tests the branch: isinstance(arg, Vector) and arg.item == (1,)
-        # Verify that Vector([4.]) has item == (1,)
-        v_scalar3 = Vector([4.])
-        self.assertEqual(v_scalar3.item, (1,))
-        p_itdiv_vec2 = Polynomial([8., 16.])
-        # This should hit the branch at line 456-457
-        p_itdiv_vec2 /= v_scalar3
-        self.assertAlmostEqual(p_itdiv_vec2.values[0], 2., places=10)
-        self.assertAlmostEqual(p_itdiv_vec2.values[1], 4., places=10)
-
-        # Test eval with order 0 and tail (drank > 0) - line 577
-        # Create a constant polynomial with drank=1
-        # For drank=1, values shape should be (..., 1, d) where d is denominator size
-        # Create a Vector with drank=1 first
-        v_const_drank = Vector(np.array([[5.]]), drank=1)  # shape (), numer (1,), denom (1,)
-        p_const_drank = Polynomial(v_const_drank)
-        result_const_drank = p_const_drank.eval(10., recursive=False)
-        self.assertEqual(type(result_const_drank), Scalar)
-        self.assertAlmostEqual(result_const_drank.values, 5., places=10)
-
-        # Test eval with order 0, derivative with tail (drank > 0) - line 589
-        v_const_deriv_drank = Vector(np.array([[7.]]), drank=1)
-        p_const_deriv_drank = Polynomial(v_const_deriv_drank)
-        v_deriv_drank = Vector(np.array([[0.]]), drank=1)
-        p_deriv_drank = Polynomial(v_deriv_drank)
-        p_const_deriv_drank.insert_deriv('t', p_deriv_drank)
-        result_const_deriv_drank = p_const_deriv_drank.eval(20., recursive=True)
-        self.assertEqual(type(result_const_deriv_drank), Scalar)
-        self.assertAlmostEqual(result_const_deriv_drank.values, 7., places=10)
-        self.assertTrue(hasattr(result_const_deriv_drank, 'd_dt'))
-
-        # Test eval with order 0, nested derivatives with tail (drank > 0) - lines 595-605
-        # This tests the full nested derivative conversion path
-        v_const_nested_drank = Vector(np.array([[9.]]), drank=1)
-        p_const_nested_drank = Polynomial(v_const_nested_drank)
-        v_deriv_nested = Vector(np.array([[0.]]), drank=1)
-        p_deriv_nested = Polynomial(v_deriv_nested)
-        # Test both branches: nested derivative that is constant (order 0) and one that is not
-        # First, test nested derivative that is constant (order 0) with tail
-        v_deriv_nested2 = Vector(np.array([[1.]]), drank=1)
-        p_deriv_nested2 = Polynomial(v_deriv_nested2)
-        p_deriv_nested.insert_deriv('s', p_deriv_nested2)
-        p_const_nested_drank.insert_deriv('t', p_deriv_nested)
-        result_const_nested_drank = p_const_nested_drank.eval(30., recursive=True)
-        self.assertEqual(type(result_const_nested_drank), Scalar)
-        self.assertAlmostEqual(result_const_nested_drank.values, 9., places=10)
-        self.assertTrue(hasattr(result_const_nested_drank, 'd_dt'))
-        # The nested derivative 's' should be converted to a Scalar
-        self.assertEqual(type(result_const_nested_drank.d_dt), Scalar)
-
-        # Also test nested derivative that is constant (order 0) with no tail (drank=0) - line 601
-        # This tests the else branch when dvalue_tail is empty
-        v_const_nested_drank2 = Vector(np.array([[11.]]), drank=1)
-        p_const_nested_drank2 = Polynomial(v_const_nested_drank2)
-        v_deriv_nested3 = Vector(np.array([[0.]]), drank=1)
-        p_deriv_nested3 = Polynomial(v_deriv_nested3)
-        # Create a nested derivative that is constant with no tail (drank=0)
-        p_deriv_nested5 = Polynomial([1.])  # constant, no tail
-        p_deriv_nested3.insert_deriv('s', p_deriv_nested5)
-        p_const_nested_drank2.insert_deriv('t', p_deriv_nested3)
-        result_const_nested_drank2 = p_const_nested_drank2.eval(40., recursive=True)
-        self.assertEqual(type(result_const_nested_drank2), Scalar)
-        self.assertAlmostEqual(result_const_nested_drank2.values, 11., places=10)
-        self.assertTrue(hasattr(result_const_nested_drank2, 'd_dt'))
-
-        # Note: Testing nested derivative that is NOT constant (order > 0) - line 605
-        # This would require a constant polynomial to have a non-constant nested derivative,
-        # but due to shape constraints, this might not be possible. The else branch at line 605
-        # handles this case, but it may be unreachable in practice.
-        # However, we can test it by creating a derivative that evaluates to a non-constant result
-        # Actually, this is complex and might not be testable. The code path exists for defensive purposes.
-
-        # Test eval with order 0, derivative that is NOT constant (line 610)
-        # This is defensive code for a case that shouldn't happen for constant polynomials
-        # But we can test it by creating a constant polynomial with a non-constant derivative
-        # Actually, this might be impossible due to shape constraints, but let's try
-        # If the polynomial is constant (order 0), its derivative should also be constant
-        # But the code has a defensive else branch. Let's see if we can trigger it.
-        # Actually, I think this line might be unreachable defensive code, but let's try
-        # to create a case where a constant polynomial has a non-constant derivative
-        # This would require the derivative to have a different shape, which might not be valid
-        # For now, let's note that this might be unreachable defensive code
-
-        # Test roots with scalar mask True (line 678)
-        p_mask_true2 = Polynomial([1., 2.], mask=True)
-        roots_mask_true2 = p_mask_true2.roots()
-        # After sort(), masked values become inf, so check for inf or mask
-        self.assertTrue(np.all(~np.isfinite(roots_mask_true2.values)) or np.all(roots_mask_true2.mask))
-
-        # Test roots with scalar mask False (line 680)
-        p_mask_false2 = Polynomial([1., 2.], mask=False)
-        roots_mask_false2 = p_mask_false2.roots()
-        # Should have no mask
-        if isinstance(roots_mask_false2.mask, np.ndarray):
-            self.assertFalse(np.any(roots_mask_false2.mask))
-        else:
-            self.assertFalse(roots_mask_false2.mask)
-
-        # Test roots with all_zeros case (lines 693-694)
-        # Create polynomial where all coefficients are zero
-        p_all_zeros2 = Polynomial([0., 0., 0.])
-        roots_all_zeros2 = p_all_zeros2.roots()
-        # Should handle all zeros case - the code sets leading coefficient to 1 and masks
-        self.assertEqual(roots_all_zeros2.shape, (2,))
-        # The all_zeros case should be masked (code sets poly_mask |= all_zeros)
-        # After sort(), masked values become inf, so check for inf or mask
-        if isinstance(roots_all_zeros2.mask, np.ndarray):
-            # Check that mask is set (all True or all inf)
-            self.assertTrue(np.all(roots_all_zeros2.mask) or np.all(~np.isfinite(roots_all_zeros2.values)))
-        else:
-            # Scalar mask case
-            self.assertTrue(roots_all_zeros2.mask or not np.any(np.isfinite(roots_all_zeros2.values)))
-
-        # Test roots with array shifts and mask_indices (lines 743-752)
-        # Create array where some elements need different numbers of shifts
-        # This tests the array case (shift_shape is not empty)
-        coeffs_array_shifts2 = np.array([
-            [[0., 0., 1., 2.], [0., 1., 2., 3.]],  # First needs 2 shifts, second needs 1 shift
-            [[1., 2., 3., 4.], [0., 0., 0., 1.]]   # First needs 0 shifts, second needs 3 shifts
-        ])
-        p_array_shifts2 = Polynomial(coeffs_array_shifts2)
-        roots_array_shifts2 = p_array_shifts2.roots()
-        # Should handle array shifts correctly
-        self.assertEqual(roots_array_shifts2.shape, (3, 2, 2))
-        # The mask_indices code path should execute when total_shifts.size > 0
-        # and len(mask_indices) > 0
-        # Line 743: if shift_shape: (array case)
-        # Line 744: if total_shifts.size > 0:
-        # Line 748: if len(mask_indices) > 0: (this should always be True for np.where results)
-
-        # Also test case where some elements have shifts but we need to ensure mask_indices is hit
-        # The code at line 748 checks if len(mask_indices) > 0, which should always be true
-        # for np.where() results, but the else branch might be unreachable
-
-        # Test roots duplicate detection scalar case (lines 768-772)
-        # Create polynomial with duplicate roots in scalar case
-        p_dup_scalar2 = Polynomial([1., -4., 4.])  # (x-2)^2, duplicate root at 2
-        roots_dup_scalar2 = p_dup_scalar2.roots()
-        # Should have duplicate masked (becomes inf after sort)
-        # In scalar case, the code checks if root_values[k] == root_values[k-1] and not root_mask
-        # If true, it sets root_mask = True and breaks
-        self.assertTrue(np.any(~np.isfinite(roots_dup_scalar2.values)) or
-                       (isinstance(roots_dup_scalar2.mask, bool) and roots_dup_scalar2.mask))
-
-        # Test roots with derivatives (lines 785-789)
-        # This tests the code path for adding derivatives to roots
-        # Use a linear polynomial for simplicity: x + 2 = 0, root at -2
-        # Derivative of polynomial: 1 (constant, nonzero at root)
-        # Derivative of polynomial w.r.t. t: some constant
-        p_roots_deriv3 = Polynomial([1., 2.])  # x + 2
-        p_roots_deriv3.insert_deriv('t', Polynomial([0., 1.]))  # derivative w.r.t. t: 1
-        roots_with_deriv3 = p_roots_deriv3.roots(recursive=True)
-        # The code path for adding derivatives (lines 785-789) should execute
-        # The derivative calculation: deriv = -value.eval(roots) / self.deriv().eval(roots)
-        # = -1 / 1 = -1
-        self.assertEqual(roots_with_deriv3.shape, (1,))
-        # Derivatives should be added
-        self.assertTrue(hasattr(roots_with_deriv3, 'd_dt'))
-        self.assertAlmostEqual(roots_with_deriv3.d_dt.values[0], -1., places=10)
-
-##########################################################################################
